requirements.txt
Flask_Cors==3.0.9
itsdangerous==1.1.0
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.4.4
Flask==1.1.2
Flask_Migrate==2.5.3
Flask_Admin==1.5.6
cryptography==2.8
flask_marshmallow==0.14.0
marshmallow==3.8.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
Flask_Script==2.0.6
flask_praetorian==1.0.0
python-dotenv==0.14.0
